+++ examples/clients.py
@@ -29,15 +29,15 @@ for client in clients:
 
 # This example iterates over ALL clients. It loads the clients gradually. In
 # pages of 30 clients.
-clients_iterator = pybillomat.ClientsIterator(conn = conn, per_page = 30)
+clients_iterator = pybillomat.ClientsIterator(conn = conn)
 clients_iterator.search()
 for client in clients_iterator:
     assert isinstance(client, pybillomat.Client)
     print client.name
 
 
-# Iterate over the first 10 clients (5 per page)
-clients_iterator = pybillomat.ClientsIterator(conn = conn, per_page = 5)
+# Iterate over the first 10 clients
+clients_iterator = pybillomat.ClientsIterator(conn = conn, per_page = 10)
 clients_iterator.search()
 for client in clients_iterator[:10]:
     assert isinstance(client, pybillomat.Client)