+++ pybillomat/__init__.py
@@ -5,6 +5,18 @@ Python-Billomat - Billomat API Client Library
 """
 
 from http import Connection
-from clients import Client, Clients, ClientsIterator
-from invoices import Invoice, Invoices, InvoicesIterator
-from clients_properties import ClientsProperty, ClientsProperties
+from clients import (
+    Client,
+    Clients,
+    ClientsIterator
+)
+from invoices import (
+    Invoice,
+    Invoices,
+    InvoicesIterator
+)
+from clients_properties import (
+    ClientsProperty,
+    ClientsProperties,
+    ClientsPropertiesIterator
+)